make object deleter replacable in the copy manager
